[Chore] ID type 수정

+++ src/main/java/groom/him/domain/search/models/entity/SearchEntity.java
@@ -24,7 +24,7 @@ public class SearchEntity extends RegisterDateFields {
     @Id
     @Column(name = "search_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer searchId;
+    private Long searchId;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private MemberEntity member;
